feat: unified settings page with secondary sidebar navigation
Consolidate Profile, License, SSO, and SMTP pages into a single
/settings route with a secondary sidebar for sub-navigation. This
reduces main sidebar clutter and provides a clear mental model
separating personal vs admin settings.

- Add Settings layout page with SettingsSidebar component
- Split Profile.tsx monolith into 8 independent section pages
- Nested React Router routes with lazy loading per section
- Admin-only and self-hosted-only filtering in secondary nav
- Responsive: vertical sidebar on desktop, horizontal pills on mobile
- Update all internal navigation links (/license -> /settings/license)
- Add settings navigation E2E tests
- Update existing profile and member management E2E tests
