+++ apps/app/public/locales/en/sidebar.json
@@ -8,6 +8,7 @@
   "users": "Users",
   "alerts": "Alerts",
   "profile": "Profile",
+  "settings": "Settings",
   "license": "License",
   "helpSupport": "Help & Support",
   "server": "Server",
