+++ apps/app/public/locales/fr/sidebar.json
@@ -8,6 +8,7 @@
   "users": "Utilisateurs",
   "alerts": "Alertes",
   "profile": "Profil",
+  "settings": "Paramètres",
   "license": "Licence",
   "helpSupport": "Aide et support",
   "server": "Serveur",
