+++ apps/app/src/components/UpgradePrompt.tsx
@@ -92,7 +92,7 @@ export function UpgradePrompt({
             <>
               <Button
                 className="flex-1 bg-gradient-button hover:bg-gradient-button-hover text-white"
-                onClick={() => navigate("/license")}
+                onClick={() => navigate("/settings/license")}
               >
                 Activate License
               </Button>
