+++ apps/app/src/pages/settings/FeedbackSection.tsx
@@ -1,0 +1,11 @@
+import { FeedbackForm } from "@/components/FeedbackForm";
+
+const FeedbackSection = () => {
+  return (
+    <div className="space-y-6">
+      <FeedbackForm />
+    </div>
+  );
+};
+
+export default FeedbackSection;
