+++ apps/app/src/pages/settings/PlansSection.tsx
@@ -1,0 +1,15 @@
+import { PlansSummaryTab } from "@/components/profile";
+
+import { useUser } from "@/hooks/ui/useUser";
+
+const PlansSection = () => {
+  const { userPlan } = useUser();
+
+  return (
+    <div className="space-y-6">
+      <PlansSummaryTab currentPlan={userPlan} />
+    </div>
+  );
+};
+
+export default PlansSection;
